requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
WTForms==3.1.0
PyMySQL==1.1.0
cryptography==41.0.7
reportlab==4.0.6
WeasyPrint==60.2
openpyxl==3.1.2
python-dotenv==1.0.0
gunicorn==21.2.0
bcrypt==4.0.1
requests==2.31.0
Pillow==10.0.1